implement add entity, added design

+++ src/app/views/admin-panel/admin-panel.component.ts
@@ -1,6 +1,8 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { UserService } from 'src/app/services/user.service';
+import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../cmps/confirm-dialog/confirm-dialog.component';
 import { RestaurantService } from 'src/app/services/restaurant.service';
 
 @Component({
@@ -12,23 +14,26 @@ export class AdminPanelComponent implements OnInit {
   data: any[] = [];
   user: any = null;
   entityData: any[] = [];
+  entityType: string = '';
   @Output() loggedInUser = new EventEmitter();
   ngOnInit() {}
   constructor(
     private http: HttpClient,
-    private restaurantService: RestaurantService
+    private restaurantService: RestaurantService,
+    private router: Router,
+    public dialog: MatDialog
   ) {
     this.getAllData();
   }
   getAllData() {
     this.entityData = [];
+    this.entityType = '';
     this.http
       .get('http://localhost:4000/api/v1/admin/', {
         withCredentials: true,
       })
       .subscribe({
         next: (response: any) => {
-          console.log('Server Response: ', response.data);
           this.data = response.data;
           this.tableData[0].count = this.data[0].length;
           this.tableData[1].count = this.data[1].length;
@@ -36,6 +41,9 @@ export class AdminPanelComponent implements OnInit {
         },
         error: (error: any) => {
           console.error('Server error', error.error);
+          if (error.status === 404) {
+            this.router.navigate(['/']);
+          }
         },
       });
   }
@@ -46,7 +54,7 @@ export class AdminPanelComponent implements OnInit {
       })
       .subscribe({
         next: (response: any) => {
-          console.log('Server Response: ', response.data);
+          this.entityType = entity;
           this.entityData = response.data;
         },
         error: (error: any) => {
@@ -58,7 +66,24 @@ export class AdminPanelComponent implements OnInit {
   summaryColumns: string[] = ['name', 'count'];
 
   deleteItem(item: any) {
-    this.restaurantService.remove(item._id);
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.restaurantService.remove(item._id);
+      }
+    });
+  }
+
+  onSelectionChange(event: any) {
+    const selectedValue = event.value;
+    if (selectedValue === 'all') {
+      this.getAllData();
+    } else {
+      this.handleData(selectedValue);
+    }
   }
   tableData = [
     { name: 'chef', count: 0 },
